Re-analyzed carbohydrate esters & amides
Ashish originally had 2 categories for carbohydrate esters and 2 categories for amides, each of which occupies a specific FTIR spectral range. I combined these 2 categories of each compound into 1 category so that there's only spectral area for carbohydrate esters and spectral area amides rather than spectral area for 2 categories of carbohydrate esters and spectral area for 2 categories of amides.

Ashish told me that combining spectral area like that is a bad idea. So, I re-analyzed these 2 functional groups in their original categories using factorial ANOVAs and Tukey's, and created boxplots for them.

+++ boxplots.py
@@ -504,7 +504,8 @@ def plotBoxPlot(enzyme, fileName):
         py.legend(handles=patches, labels=legendLabels)
 
     figPath = tukeyFolder/enzyme/figName
-    py.savefig(figPath, bbox_inches="tight", pad_inches=0.04)
+    if os.path.exists(figPath) is False:
+        py.savefig(figPath, bbox_inches="tight", pad_inches=0.04)
     return
